+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- empty list handling in datenkompass
+
 ### Security
 
 ## [1.1.0] - 2025-09-11